requirements.txt
Flask==2.3.3
tensorflow==2.13.0
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.0
pickle-mixin==1.0.2
python-dotenv==1.0.0
flask-cors==4.0.0
